Added timeouts for writing and reading from zabbix
Previously only a connection timeout was defined, with a default value
of 15s and configurable with an optional parameter in the NewSender
function.

This commits adds timeouts for writing data and reading data from
Zabbix.
When Zabbix is loaded, responses could take several seconds, especially
when LLDs are sent.

To be able to configure those timeouts a new function
``NewSenderTimeout`` is created, also the timeout option in
``NewSender`` is removed.

Also code has been simplified using ``DialTimeout``